added: component integration
