+++ src/app/pages/products/components/index.ts
@@ -1,1 +1,4 @@
 export { default as SmartSolutions } from "./SmartSolutions";
+export { default as LearnMoreCRM } from "./LearnMoreCRM";
+export { default as ScaleYourBusiness } from "./ScaleYourBusiness";
+export { default as EmpowerBusiness } from "./EmpowerBusiness";
